requirements/devel.txt
-r base.txt

django-debug-toolbar
django-cache-panel
